transform/src/requirements.txt
google-api-core==2.7.3
google-cloud-bigquery==3.1.0
google-cloud-storage==2.3.0
pandas==1.3.3